сделал новый модуль, импортировал в main
